+++ custom_plugin_example/lib/custom_log.dart
@@ -4,8 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_ume/flutter_ume.dart';
 
 class CustomLog implements Pluggable {
+
   @override
-  Widget buildWidget(BuildContext context) => _buildLogPanel(context);
+  Widget? buildWidget(BuildContext? context) => _buildLogPanel(context);
 
   @override
   ImageProvider<Object> get iconImageProvider => MemoryImage(base64Decode(
@@ -26,11 +27,12 @@ class CustomLog implements Pluggable {
   static List<String> get logList => _logList;
 
   static void log(String info) {
-    _logList.add(info ?? '');
+    _logList.add(info);
     print(info);
   }
 
-  Widget _buildLogPanel(BuildContext context) {
+  Widget _buildLogPanel(BuildContext? context) {
+    if (context == null) return Container();
     return Align(
       alignment: Alignment.bottomCenter,
       child: SizedBox(
@@ -41,7 +43,7 @@ class CustomLog implements Pluggable {
 }
 
 class LogViewerPage extends StatefulWidget {
-  LogViewerPage({Key key}) : super(key: key);
+  LogViewerPage({Key? key}) : super(key: key);
 
   @override
   _LogViewerPageState createState() => _LogViewerPageState();