+++ kits/flutter_ume_kit_channel_monitor/lib/flutter_ume_kit_channel_monitor.dart
@@ -1,0 +1,3 @@
+library flutter_ume_kit_channel_monitor;
+
+export 'src/channel_plugin.dart';