+++ kits/flutter_ume_kit_console/lib/console/console_print.dart
@@ -1,0 +1,6 @@
+import 'console_manager.dart';
+
+/// Print the message to the console.
+void consolePrint(String message) {
+  ConsoleManager.streamController!.add(message);
+}