+++ kits/flutter_ume_kit_dio/lib/src/widgets/pluggable_state.dart
@@ -229,6 +229,20 @@ class _ResponseCardState extends State<_ResponseCard> {
   /// The [Uri] that the [_request] requested.
   Uri get _requestUri => _request.uri;
 
+  String? get _requestHeadersBuilder {
+    final Map<String, List<String>> map = _request.headers.map(
+      (key, value) => MapEntry(
+        key,
+        value is Iterable ? value.map((v) => v.toString()).toList() : [value],
+      ),
+    );
+    final Headers headers = Headers.fromMap(map);
+    if (headers.isEmpty) {
+      return null;
+    }
+    return '$headers';
+  }
+
   /// Data for the [_request].
   String? get _requestDataBuilder {
     if (_request.data is Map) {
@@ -238,11 +252,22 @@ class _ResponseCardState extends State<_ResponseCard> {
   }
 
   /// Data for the [_response].
-  String get _responseDataBuilder {
+  String? get _responseDataBuilder {
+    final data = _response.data;
+    if (data == null) {
+      return null;
+    }
     if (_response.data is Map) {
       return _encoder.convert(_response.data);
     }
     return _response.data.toString();
+  }
+
+  String? get _responseHeadersBuilder {
+    if (_response.headers.isEmpty) {
+      return null;
+    }
+    return '${_response.headers}';
   }
 
   Widget _detailButton(BuildContext context) {
@@ -295,17 +320,26 @@ class _ResponseCardState extends State<_ResponseCard> {
         if (!value) {
           return const SizedBox.shrink();
         }
-        return Container(
+        return Padding(
           padding: const EdgeInsets.symmetric(vertical: 8),
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
-              if (_requestDataBuilder != null)
-                _TagText(tag: 'Request data', content: _requestDataBuilder!),
-              _TagText(tag: 'Response body', content: _responseDataBuilder),
+              _TagText(
+                tag: 'Request headers',
+                content: _requestHeadersBuilder,
+              ),
+              _TagText(
+                tag: 'Request data',
+                content: _requestDataBuilder,
+              ),
+              _TagText(
+                tag: 'Response body',
+                content: _responseDataBuilder,
+              ),
               _TagText(
                 tag: 'Response headers',
-                content: '\n${_response.headers}',
+                content: _responseHeadersBuilder,
               ),
             ],
           ),
@@ -327,7 +361,11 @@ class _ResponseCardState extends State<_ResponseCard> {
           children: <Widget>[
             _infoContent(context),
             const SizedBox(height: 10),
-            _TagText(tag: 'Uri', content: '$_requestUri'),
+            _TagText(
+              tag: 'Uri',
+              content: '$_requestUri',
+              shouldStartFromNewLine: false,
+            ),
             _detailedContent(context),
           ],
         ),
@@ -340,13 +378,13 @@ class _TagText extends StatelessWidget {
   const _TagText({
     Key? key,
     required this.tag,
-    required this.content,
-    this.selectable = true,
+    this.content,
+    this.shouldStartFromNewLine = true,
   }) : super(key: key);
 
   final String tag;
-  final String content;
-  final bool selectable;
+  final String? content;
+  final bool shouldStartFromNewLine;
 
   TextSpan get span {
     return TextSpan(
@@ -355,28 +393,22 @@ class _TagText extends StatelessWidget {
           text: '$tag: ',
           style: const TextStyle(fontWeight: FontWeight.bold),
         ),
-        TextSpan(text: content.notBreak),
+        if (shouldStartFromNewLine) TextSpan(text: '\n'),
+        TextSpan(text: content!),
       ],
     );
   }
 
   @override
   Widget build(BuildContext context) {
-    Widget text;
-    if (selectable) {
-      text = SelectableText.rich(span);
-    } else {
-      text = Text.rich(span);
+    if (content == null) {
+      return const SizedBox.shrink();
     }
     return Padding(
       padding: const EdgeInsets.symmetric(vertical: 4),
-      child: text,
+      child: SelectableText.rich(span),
     );
   }
-}
-
-extension _StringExtension on String {
-  String get notBreak => Characters(this).toList().join('\u{200B}');
 }
 
 extension _DateTimeExtension on DateTime {