+++ lib/core/ui/dragable_widget.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
-typedef bool CanAccept(int oldIndex, int newIndex);
+typedef CanAccept = bool Function(int oldIndex, int newIndex);
 
-typedef void DragCompletion<T>(List<T?>? data);
+typedef DragCompletion<T> = void Function(List<T?>? data);
 
-typedef Widget DataWidgetBuilder<T>(BuildContext context, T data);
+typedef DataWidgetBuilder<T> = Widget Function(BuildContext context, T data);
 
 class DragableGridView<T> extends StatefulWidget {
   final DataWidgetBuilder<T> itemBuilder;
@@ -24,10 +24,7 @@ class DragableGridView<T> extends StatefulWidget {
     this.dragCompletion,
     required this.itemBuilder,
     required this.canAccept,
-  })  : assert(itemBuilder != null),
-        assert(canAccept != null),
-        assert(dataList != null && dataList.length >= 0),
-        super(key: key);
+  }) : super(key: key);
 
   @override
   State<StatefulWidget> createState() => _DragableGridViewState<T>();