+++ lib/core/ui/menu_page.dart
@@ -1,10 +1,8 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_ume/core/ui/icon_cache.dart';
 import 'package:flutter_ume/core/pluggable_message_service.dart';
 import 'package:flutter_ume/core/red_dot.dart';
 import 'package:flutter_ume/core/store_manager.dart';
-import 'package:flutter_ume/core/plugin_manager.dart';
 import 'package:flutter_ume/flutter_ume.dart';
 import 'dragable_widget.dart';
 import 'package:flutter_ume/core/ui/panel_action_define.dart';
@@ -59,7 +57,7 @@ class _MenuPageState extends State<MenuPage>
 
   void _saveData(List<Pluggable?> data) {
     List l = data.map((f) => f!.name).toList();
-    if (l == null || l.isEmpty) {
+    if (l.isEmpty) {
       return;
     }
     Future.delayed(Duration(milliseconds: 500), () {