+++ kits/flutter_ume_kit_console/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.1.0
+
+* 增加 `consolePrint` 方法，用于适配其他日志库。
+
+* Add `consolePrint` method to adapt other log libraries.
+
+## 1.0.0
+
+* 正式版
+
+* Normal version.
+
+## 1.0.0-dev.0
+
+* 适配 Flutter 3
+
+* Adapt Flutter 3
+
 ## 0.3.0
 
 * 更新版本号